Integration of spatialite-simple
