+++ dat/mark-rows.tsv
@@ -1,0 +1,42 @@
+1	HE	East Fairway Buoy	PHM	50.877333	-1.498000	Red	Can
+2	1	Black Rock	SHM	50.708833	-1.507167	Green	Cone
+3	4	The English Summer Berry Co.	RM	50.714333	-1.473667	Yellow	Sphere
+4	1G	SW Shingles	PHM	50.654833	-1.691667	Red	Pillar
+5	1M	Warden	SHM	50.691333	-1.559167	Green	Can
+6	2	Charles Stanley	RM	50.714333	-1.490000	Yellow	Sphere
+7	3	Folkboats Forever	RM	50.707833	-1.507167	Yellow	Sphere
+8	#	Towers	RM	50.710500	-1.487667	Yellow	Pillar
+9	WH	West Harbour Entrance	SHM	50.706833	-1.501167	Green	Cone
+10	5	Caldwells	RM	50.719167	-1.458167	Yellow	Cylinder
+11	6	Hamstead Ledge	SHM	50.731000	-1.436333	Green	Cone
+12	7	Alitex	RM	50.716167	-1.469333	Yellow	Can
+13	8	Bouldnor	RM	50.712167	-1.478167	Yellow	Cylinder
+14	9	Royal Solent YC ODM	RM	50.712000	-1.495000	Yellow	Cylinder
+15	1N	NE Shingles	ECard	50.699333	-1.556667	BYB	Pillar
+16	A	Hurst	RM	50.713167	-1.542167	Yellow	Sphere
+17	B	Pennington	RM	50.723000	-1.525667	Yellow	Cylinder
+18	C	Oxey	RM	50.730500	-1.514333	Yellow	Cylinder
+19	D	Royal Lymington	RM	50.736167	-1.501667	Yellow	Cylinder
+20	E 	Pylewell	RM	50.743000	-1.490500	Yellow	Cylinder
+21	F	Berthon	RM	50.736667	-1.487000	Yellow	Sphere
+22	G	Tanners	RM	50.746500	-1.474500	Yellow	Cylinder
+23	H	Jackson	RM	50.738333	-1.469333	Yellow	Cylinder
+24	J	Sowley	RM	50.751833	-1.455500	Yellow	Sphere
+25	K	Quod Possumus!	RM	50.747167	-1.434833	Yellow	Sphere
+26	R	Durns	RM	50.757167	-1.431500	Yellow	Sphere
+27	T	Lymington Bank	PHM	50.718333	-1.514167	Red	Can
+28	V	Victor	RM	50.721667	-1.521667	Yellow	Can
+29	X	Solent Bank	PHM	50.737167	-1.456167	Red	Can
+30	0	Sconse	NCard	50.708667	-1.523833	BBY	Pillar
+31	3A	West Lepe	PHM	50.753833	-1.401500	Red	Can
+32	3D	Salt Mead	SHM	50.742000	-1.384000	Green	Cone
+33	3H	East Lepe	PHM	50.765500	-1.351167	Red	Can
+34	3L	Gurnard Ledge	SHM	50.758500	-1.343167	Green	Cone
+35	JB	Jack in the Basket	PHM	50.738000	-1.509500	Red	Post
+36	11	Delphie	RM	50.708000	-1.499000	Yellow	Sphere
+37	14	Stone Pier	RM	50.713667	-1.467500	Yellow	Sphere
+38	15	Hallmark	RM	50.710500	-1.478333	Yellow	Sphere
+39	16	St Swithins	RM	50.708667	-1.483167	Yellow	Sphere
+40	17	Camel	RM	50.709500	-1.486167	Yellow	Sphere
+41	18	Mount	RM	50.707333	-1.490500	Yellow	Sphere
+42	19	Dinghy ODM	RM	50.707333	-1.497000	Yellow	Sphere
