+++ deepblue.cabal
@@ -34,9 +34,7 @@ library
                            , Deepblue.Graphics.Status
                            , Deepblue.Graphics.World
                            , Deepblue.Graphics.Interact
-                           , Deepblue.Database.GIS
-                           , Deepblue.Database.GIS.Schema
-        other-modules:       Deepblue.Database.GIS.Internal
+        other-modules:       
                            
         build-depends:       base
                            , geodetics
@@ -46,8 +44,9 @@ library
                            , sort
                            , containers
                            , streamly
-                           , direct-sqlite
+                           --, direct-sqlite
                            , formatting
+                           , spatialite-simple
                            --, hnetcdf
                            --, random
                            --, mtl
@@ -89,6 +88,22 @@ executable loadgeo
                      , deepblue
 
   hs-source-dirs:      utl
+  default-language:    Haskell2010
+  ghc-options: -O3 -Wall -threaded
+  extra-libraries: sqlite3
+  extra-lib-dirs: /usr/local/opt/sqlite3/lib /usr/local/lib
+
+executable check
+  main-is:             Main.hs
+  -- other-modules:    
+  -- other-extensions:    
+  build-depends:       base
+                     , cmdargs
+                     , deepblue
+                     , spatialite-simple
+                     , text
+
+  hs-source-dirs:      tst
   default-language:    Haskell2010
   ghc-options: -O3 -Wall -threaded
   extra-libraries: sqlite3
