+++ sql/marks.sql
@@ -1,0 +1,4 @@
+create table if not exists marks (id INTEGER primary key autoincrement, label TEXT, description TEXT, marktype TEXT, lat FLOAT, lon FLOAT, color TEXT, shape TEXT);
+-- load data from file
+-- add postition column
+-- update position from lon, lat
