--- src/Deepblue/Database/GIS.hs
@@ -1,7 +1,0 @@
-module Deepblue.Database.GIS 
-    ( module Deepblue.Database.GIS.Schema
-    , module Deepblue.Database.GIS.Internal
-    ) where
-
-import Deepblue.Database.GIS.Schema
-import Deepblue.Database.GIS.Internal
