--- stack-ghc-8.6.5.yaml
@@ -1,55 +1,0 @@
-resolver: lts-14.7
-
-# User packages to be built.
-# Various formats can be used as shown in the example below.
-#
-# packages:
-# - some-directory
-# - https://example.com/foo/bar/baz-0.0.2.tar.gz
-#   subdirs:
-#   - auto-update
-#   - wai
-packages:
-  - .
-
-# Dependency packages to be pulled from upstream that are not in the resolver.
-# These entries can reference officially published versions as well as
-# forks / in-progress versions pinned to a git hash. For example:
-#
-extra-deps:
-  - git: https://github.com/sebeaumont/geodetics.git
-    commit: 8e7f7b1cc165462f28f9b5c31b85beedb1b99188
-  - time-1.9.3
-  - directory-1.3.4.0
-  - process-1.6.5.1
-  - unix-2.7.2.2
-  - gloss-1.13.1.1
-  - gloss-rendering-1.13.1.1
-  
-# this for netcdf and grenade -- TODO must bump dependencies therein...
-allow-newer: true
-
-# Override default flag values for local packages and extra-deps
-# flags: {}
-
-# Extra package databases containing global packages
-# extra-package-dbs: []
-
-# Control whether we use the GHC we find on the path
-#system-ghc: true
-
-#
-# Require a specific version of stack, using version ranges
-# require-stack-version: -any # Default
-# require-stack-version: ">=2.1"
-#
-# Override the architecture used by stack, especially useful on Windows
-# arch: i386
-# arch: x86_64
-#
-# Extra directories used by stack for building
-# extra-include-dirs: [/path/to/dir]
-# extra-lib-dirs: [/path/to/dir]
-#
-# Allow a newer minor version of GHC than the snapshot specifies
-# compiler-check: newer-minor
