--- stack-ghc-8.6.5.yaml.lock
@@ -1,68 +1,0 @@
-# This file was autogenerated by Stack.
-# You should not edit this file by hand.
-# For more information, please see the documentation at:
-#   https://docs.haskellstack.org/en/stable/lock_files
-
-packages:
-- completed:
-    cabal-file:
-      size: 2387
-      sha256: 699c10cd8d69222b125d054cdc8e2bcee7cb0558c78fa1b4fa06b2d3ecf29f5c
-    name: geodetics
-    version: 0.1.0
-    git: https://github.com/sebeaumont/geodetics.git
-    pantry-tree:
-      size: 1186
-      sha256: d26a259705d2a96fa8efd30e9061f1572346e3ab9a3a50886184a49341738e1e
-    commit: 8e7f7b1cc165462f28f9b5c31b85beedb1b99188
-  original:
-    git: https://github.com/sebeaumont/geodetics.git
-    commit: 8e7f7b1cc165462f28f9b5c31b85beedb1b99188
-- completed:
-    hackage: time-1.9.3@sha256:8f1b5448722a12a952248b356c9eb366e351226543d9086a2da71270522d5f45,5679
-    pantry-tree:
-      size: 6558
-      sha256: a1043c1719491764f0fa37a1fd70d9451080548a41632fee88d8e1b8db4942d6
-  original:
-    hackage: time-1.9.3
-- completed:
-    hackage: directory-1.3.4.0@sha256:500019f04494324d1df16cf83eefeb3f809b2b20b32a32ccd755ee0439c18bfd,2829
-    pantry-tree:
-      size: 3365
-      sha256: 00c09e0c014d29ebfb921b64c1459e61a0ad6f10e70128d795246a47c06394b0
-  original:
-    hackage: directory-1.3.4.0
-- completed:
-    hackage: process-1.6.5.1@sha256:77a9afeb676357f67fe5cf1ad79aca0745fb6f7fb96b786d510af08f622643f6,2468
-    pantry-tree:
-      size: 1211
-      sha256: 19d944da6aa37944332e0726372288319852e5f72aa57dbc3516dc15e760a502
-  original:
-    hackage: process-1.6.5.1
-- completed:
-    hackage: unix-2.7.2.2@sha256:e69269a17b9fa26cb77f3f55e86c39d0a4940ccfa0c4bc20826919d2572076ad,3496
-    pantry-tree:
-      size: 3536
-      sha256: c355f7924ce67e5bf8f20767462af18f09b8c0d1f7161117221cbb94c15deee3
-  original:
-    hackage: unix-2.7.2.2
-- completed:
-    hackage: gloss-1.13.1.1@sha256:49d2d0c7c4a0a5b55a344a257b225b6b3fdd811eaa9c960387a353e094496333,4065
-    pantry-tree:
-      size: 3807
-      sha256: 34a70cb69083d60c18630b0db819d252568d1a67a4c8b05e14c69ff3e40cfdae
-  original:
-    hackage: gloss-1.13.1.1
-- completed:
-    hackage: gloss-rendering-1.13.1.1@sha256:dc6c2d26ea5e19ee4b24ae17d6f5e45eb77df3c34ad87c732fae7a1cd6e5b246,1650
-    pantry-tree:
-      size: 895
-      sha256: f723afd2ab1a3a7c6e128f657268c35f2bb2084a2789543619aa804fd6bfab74
-  original:
-    hackage: gloss-rendering-1.13.1.1
-snapshots:
-- completed:
-    size: 523700
-    url: https://raw.githubusercontent.com/commercialhaskell/stackage-snapshots/master/lts/14/7.yaml
-    sha256: 8e3f3c894be74d71fa4bf085e0a8baae7e4d7622d07ea31a52736b80f8b9bb1a
-  original: lts-14.7
