+++ stack.yaml
@@ -4,10 +4,19 @@ packages:
   - .
 
 extra-deps:
+  # this I contribute to
   - git: https://github.com/sebeaumont/geodetics.git
     commit: 8e7f7b1cc165462f28f9b5c31b85beedb1b99188
+  # this is mine
+  - git: https://github.com/sebeaumont/spatialite-simple.git
+    commit: f80f5d7d381c0bb474c42633eaf072a72f2a4b37
+  # this for ghc-8.8 patches
+  - git: https://github.com/sebeaumont/sqlite-simple.git
+    commit: 733fcecc5dd9336baa1f6a072c11375b1b61165c
+  # this to ensure built against correct libs 
   - git: https://github.com/sebeaumont/direct-sqlite.git
     commit: 36f3a7aa87c538a888a22d0077314fa0996ed44d
+
   - time-1.9.3
   - directory-1.3.4.0
   - process-1.6.5.1
@@ -15,7 +24,7 @@ extra-deps:
   - gloss-1.13.1.1
   - gloss-rendering-1.13.1.1
   
-  
+
 # this for netcdf and grenade -- TODO must bump dependencies therein...
 allow-newer: true
